:recycle: Extracts to Module
