+++ Cargo.toml
@@ -1,15 +1,23 @@
 [package]
-name = "url-shortener"
+name = "url_shortener"
 version = "0.0.1"
 edition = "2021"
+
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+path = "src/main.rs"
+name = "url_shortener"
+
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 serde = "1.0.162"
 surrealdb = { git = "https://github.com/surrealdb/surrealdb.git", tag = "v1.0.0-beta.9" }
-tokio = { version= "1.28.0", features = ["full"] }
-warp = "0.3.5"
+tokio = { version = "1.28.0", features = ["time", "macros", "tokio-macros"] }
+warp = { version ="0.3.5", default-features = false, features = ["tokio-rustls", "websocket"] }
 
 
 [profile.release]
