+++ tests/health_check.rs
@@ -1,0 +1,6 @@
+use url_shortener::run;
+
+#[test]
+fn dummy() {
+    assert_eq!(0, 0);
+}
